Refactor code to move quasar code to experimental quasar folder
