+++ ttnn/cpp/ttnn/operations/eltwise/binary_ng/sources.cmake
@@ -4,6 +4,5 @@
 set(TTNN_OP_ELTWISE_BINARY_NG_SRCS
     device/binary_ng_device_operation.cpp
     device/binary_ng_program_factory.cpp
-    device/binary_ng_metal_v2_factory.cpp
     device/binary_ng_utils.cpp
 )
