+++ ttnn/cpp/ttnn/operations/experimental/quasar/binary_ng/device/binary_ng_device_operation.hpp
@@ -5,6 +5,7 @@
 #pragma once
 
 #include "ttnn/device_operation.hpp"
+#include "ttnn/metal_v2_artifacts.hpp"
 #include "ttnn/operations/core/compute_kernel/compute_kernel_config.hpp"
 #include "ttnn/operations/experimental/quasar/binary_ng/types.hpp"
 #include "ttnn/operations/eltwise/unary/common/unary_op_types.hpp"
@@ -71,7 +72,30 @@ struct BinaryNgDeviceOperation {
             tensor_return_value_t& c);
     };
 
-    using program_factory_t = std::variant<ProgramFactory>;
+    // Metal 2.0 / DataflowBuffer factory. Emits a ProgramSpec + ProgramRunArgs (instead of a
+    // ProgramDescriptor) for the slice that matches_metal_v2_slice() admits: ADD, no broadcast,
+    // TILE 32x32, fully height/block-sharded L1, bf8/bf16, optional fused RELU, in-place (the
+    // ResNet50 residual config). The DFB path is arch-portable: CB-backed on Wormhole/Blackhole,
+    // overlay-backed on Quasar. `select_program_factory` routes a matching op here; every other case
+    // falls through to `ProgramFactory`. See binary_ng_metal_v2_factory.cpp.
+    struct ProgramFactoryMetalV2 {
+        static ttnn::device_operation::ProgramArtifacts create_program_artifacts(
+            const operation_attributes_t& operation_attributes,
+            const tensor_args_t& tensor_args,
+            tensor_return_value_t& c);
+    };
+
+    using program_factory_t = std::variant<ProgramFactory, ProgramFactoryMetalV2>;
+
+    // Returns ProgramFactoryMetalV2{} when the op matches the Metal 2.0 slice; ProgramFactory{}
+    // otherwise.
+    static program_factory_t select_program_factory(const operation_attributes_t&, const tensor_args_t&);
+
+    // True iff (attributes, tensor_args) match the Metal 2.0 factory's supported slice. Shared by
+    // select_program_factory and compute_program_hash (to fold the shape into the hash for matching
+    // ops only).
+    static bool matches_metal_v2_slice(const operation_attributes_t&, const tensor_args_t&);
+
     static void validate_on_program_cache_miss(const operation_attributes_t&, const tensor_args_t&);
     static void validate_on_program_cache_hit(const operation_attributes_t&, const tensor_args_t&);
     static spec_return_value_t compute_output_specs(const operation_attributes_t&, const tensor_args_t&);
